Resolve dependencies in ibc-clients crate
