+++ crates/ibc-clients/Cargo.toml
@@ -6,30 +6,24 @@ edition      = { workspace = true }
 rust-version = { workspace = true }
 license      = { workspace = true }
 repository   = { workspace = true }
+keywords     = ["blockchain", "cosmos", "ibc", "applications", "tendermint"]
 readme       = "README.md"
-keywords     = ["blockchain", "consensus", "cosmos", "ibc", "tendermint"]
 description  = """
-    `ibc-clients` exports the types, handlers, and contexts for implementing Tendermint,
-    CosmWasm, and other IBC clients.
+    `ibc-clients` provides a comprehensive set of libraries for IBC client implementations, 
+    facilitating seamless integration of IBC business logic into any blockchain system.
 """
 
+[package.metadata.docs.rs]
+all-features = true
+
 [dependencies]
-# external dependencies
-bytes = { workspace = true }
-displaydoc = { workspace = true }
-prost = { workspace = true }
+ibc-client-tendermint = { version = "0.47.0", path = "./ics07-tendermint", default-features = false }
 
-# cosmos dependencies
-tendermint = { workspace = true }
-tendermint-light-client-verifier = { workspace = true, features = ["rust-crypto"] }
-tendermint-proto = { workspace = true }
-
-# ibc dependencies
-ibc = { version = "0.47.0", path = "../ibc" }
-ibc-proto = { workspace = true, features = ["server"] }
-
-[dev-dependencies]
-env_logger = { workspace = true }
-ibc-testkit = { version = "0.47.0", path = "../ibc-testkit", default-features = false, features = ["serde"] }
-test-log = { workspace = true }
-tracing-subscriber = { workspace = true }
+# [features]
+# default = ["std"]
+# std = [
+#     "ibc-client-tendermint/std",
+# ]
+# serde = [
+#     "ibc-client-tendermint/serde",
+# ]
