+++ crates/ibc-clients/ics07-tendermint/src/consensus_state.rs
@@ -1,0 +1,13 @@
+impl ConsensusStateTrait for ConsensusState {
+    fn root(&self) -> &CommitmentRoot {
+        &self.root
+    }
+
+    fn timestamp(&self) -> Timestamp {
+        self.timestamp.into()
+    }
+
+    fn encode_vec(self) -> Vec<u8> {
+        <Self as Protobuf<Any>>::encode_vec(self)
+    }
+}
