--- crates/ibc-clients/src/ibc_client_tendermint/impls/mod.rs
@@ -1,1 +1,0 @@
-pub mod context;
