--- crates/ibc-clients/src/ibc_client_tendermint/types/mod.rs
@@ -1,6 +1,0 @@
-pub mod client_state;
-pub mod consensus_state;
-pub mod header;
-pub mod misbehaviour;
-pub mod trust_threshold;
-
